Rename files to get rid of camel/pascal case
